// src/modules/graphql/arweave.net/types/Schema.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
};
// ...
export type TransactionEdge = {
  __typename?: "TransactionEdge";
  cursor: Scalars["String"]["output"];
  node: Transaction;
};

export type PageInfo = {
  __typename?: "PageInfo";
  hasNextPage: Scalars["Boolean"]["output"];
};

export type Transaction = {
  __typename?: "Transaction";
  id: Scalars["ID"]["output"];
  anchor: Scalars["String"]["output"];
  signature: Scalars["String"]["output"];
  recipient: Scalars["String"]["output"];
  owner: Owner;
  fee: Amount;
  quantity: Amount;
  data: MetaData;
  tags: Array<Tag>;
  block?: Maybe<Block>;
  /** @deprecated Use `bundledIn` */
  parent?: Maybe<Parent>;
  bundledIn?: Maybe<Bundle>;
};

export type Parent = {
  __typename?: "Parent";
  id: Scalars["ID"]["output"];
};

export type Bundle = {
  __typename?: "Bundle";
  id: Scalars["ID"]["output"];
};

export type Block = {
  __typename?: "Block";
  id: Scalars["ID"]["output"];
  timestamp: Scalars["Int"]["output"];
  height: Scalars["Int"]["output"];
  previous: Scalars["ID"]["output"];
};

export type MetaData = {
  __typename?: "MetaData";
  size: Scalars["String"]["output"];
  type?: Maybe<Scalars["String"]["output"]>;
};

export type Amount = {
  __typename?: "Amount";
  winston: Scalars["String"]["output"];
  ar: Scalars["String"]["output"];
};

export type Owner = {
  __typename?: "Owner";
  address: Scalars["String"]["output"];
  key: Scalars["String"]["output"];
};

export type Tag = {
  __typename?: "Tag";
  name: Scalars["String"]["output"];
  value: Scalars["String"]["output"];
};